Rename several svg files
