+++ src/shared/components/plus-button/PlusButton.tsx
@@ -16,7 +16,7 @@
 
 import React, { type ButtonHTMLAttributes } from 'react';
 
-import { ReactComponent as PlusIcon } from 'static/icons/icon_nav_plus.svg';
+import { ReactComponent as PlusIcon } from 'static/icons/icon_plus_circle.svg';
 
 import styles from './PlusButton.scss';
