+++ src/shared/components/visibility-icon/VisibilityIcon.tsx
@@ -21,7 +21,7 @@ import ImageButton, {
 } from 'src/shared/components/image-button/ImageButton';
 
 import { ReactComponent as Eye } from 'static/icons/icon_eye.svg';
-import { ReactComponent as EyePartial } from 'static/icons/icon_eye_half.svg';
+import { ReactComponent as EyePartial } from 'static/icons/icon_eye_slashed.svg';
 
 import { Status } from 'src/shared/types/status';
